backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.2
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
celery==5.4.0
redis==5.1.1
werkzeug==3.0.4